backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
psycopg[binary]==3.2.9
alembic==1.13.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.136
qrcode==7.4.2
reportlab==4.2.2
pillow==10.4.0
httpx==0.27.0
pytest==8.2.2
